Revert "Revert "Revert "Finished integration and unit testing -- Successful"""
This reverts commit 8e433649fff467b952a2db86f29573ae371b1c0d.

+++ test_app/db/migrate/20210903165213_create_books.rb
@@ -3,8 +3,9 @@ class CreateBooks < ActiveRecord::Migration[6.1]
     create_table :books do |t|
       t.string :title
       t.string :author
-      t.float :price
+      t.integer :price
       t.date :published_date
+
       t.timestamps
     end
   end
